+++ NEWS.md
@@ -1,8 +1,9 @@
 # mlr3pipelines 0.2.1-9000
 
 * NULL input channels accept any kind of input
+* PipeOpTextVectorizer can now return an "integer sequence representation".
 * New PipeOps:
-- PipeOpNMF
+  - PipeOpNMF
 
 # mlr3pipelines 0.2.1