+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * NULL input channels accept any kind of input
 * state of PipeOps now reset to NULL if training fails
+* implemented as_learner.PipeOp
 * Changed PPLs:
   - fix how ppl_robustify detects whether a learner can handle factors
 * Changed PipeOps: