+++ NEWS.md
@@ -68,6 +68,9 @@
 * predict() generic for Graph
 * Migrated last vignette to "mlr3 Book"
 
+* Compact in-memory representation of R6 objects to save space when
+  saving objects via saveRDS(), serialize() etc.
+
 # mlr3pipelines 0.1.2
 
 * Work with new mlr3 version 0.1.5 (handling of character columns changed)