+++ NEWS.md
@@ -2,6 +2,15 @@
 
 * NULL input channels accept any kind of input
 * New to_dot() method added to Graphs printing a DOT representation on console
+* state of PipeOps now reset to NULL if training fails
+* implemented as_learner.PipeOp
+* Changed PPLs:
+  - fix how ppl_robustify detects whether a learner can handle factors
+* Changed PipeOps:
+  - PipeOpTextVectorizer can now return an "integer sequence representation".
+* New PipeOps:
+  - PipeOpNMF
+  - PipeOpColRoles
 
 # mlr3pipelines 0.2.1
 
@@ -59,6 +68,9 @@
 * small test fix for R 4.0 (necessary for stringsAsFactors option default change in 3.6 -> 4.0)
 * predict() generic for Graph
 * Migrated last vignette to "mlr3 Book"
+
+* Compact in-memory representation of R6 objects to save space when
+  saving objects via saveRDS(), serialize() etc.
 
 # mlr3pipelines 0.1.2